+++ r2dbc/src/test/java/org/jdbi/v3/r2dbc/QueryTest.java
@@ -14,25 +14,26 @@
  * limitations under the License.
  */
 
-package io.r2dbc.client;
+package org.jdbi.v3.r2dbc;
 
 import java.util.Collections;
 
 import io.r2dbc.spi.test.MockResult;
 import io.r2dbc.spi.test.MockStatement;
 import org.junit.jupiter.api.Test;
+import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-final class UpdateTest {
+final class QueryTest {
 
     @Test
     void add() {
         MockStatement statement = MockStatement.empty();
 
-        new Update(statement)
+        new Query(statement)
             .add();
 
         assertThat(statement.isAddCalled()).isTrue();
@@ -42,7 +43,7 @@ final class UpdateTest {
     void bind() {
         MockStatement statement = MockStatement.empty();
 
-        new Update(statement)
+        new Query(statement)
             .bind("test-identifier", "test-value");
 
         assertThat(statement.getBindings()).contains(Collections.singletonMap("test-identifier", "test-value"));
@@ -52,7 +53,7 @@ final class UpdateTest {
     void bindIndex() {
         MockStatement statement = MockStatement.empty();
 
-        new Update(statement)
+        new Query(statement)
             .bind(100, "test-value");
 
         assertThat(statement.getBindings()).contains(Collections.singletonMap(100, "test-value"));
@@ -60,19 +61,19 @@ final class UpdateTest {
 
     @Test
     void bindIndexNoValue() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Update(MockStatement.empty()).bind(100, null))
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(MockStatement.empty()).bind(100, null))
             .withMessage("value must not be null");
     }
 
     @Test
     void bindNoIdentifier() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Update(MockStatement.empty()).bind(null, new Object()))
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(MockStatement.empty()).bind(null, new Object()))
             .withMessage("identifier must not be null");
     }
 
     @Test
     void bindNoValue() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Update(MockStatement.empty()).bind("test-identifier", null))
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(MockStatement.empty()).bind("test-identifier", null))
             .withMessage("value must not be null");
     }
 
@@ -80,7 +81,7 @@ final class UpdateTest {
     void bindNull() {
         MockStatement statement = MockStatement.empty();
 
-        new Update(statement)
+        new Query(statement)
             .bindNull("test-identifier", Integer.class);
 
         assertThat(statement.getBindings()).contains(Collections.singletonMap("test-identifier", Integer.class));
@@ -88,37 +89,44 @@ final class UpdateTest {
 
     @Test
     void bindNullNoIdentifier() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Update(MockStatement.empty()).bindNull(null, Object.class))
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(MockStatement.empty()).bindNull(null, Object.class))
             .withMessage("identifier must not be null");
     }
 
     @Test
     void bindNullNoType() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Update(MockStatement.empty()).bindNull("test-identifier", null))
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(MockStatement.empty()).bindNull("test-identifier", null))
             .withMessage("type must not be null");
     }
 
     @Test
     void constructorNoStatement() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Update(null))
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(null))
             .withMessage("statement must not be null");
     }
 
     @Test
-    void execute() {
-        MockResult result = MockResult.builder()
-            .rowsUpdated(100)
-            .build();
+    void mapResult() {
+        MockResult result = MockResult.empty();
 
         MockStatement statement = MockStatement.builder()
             .result(result)
             .build();
 
-        new Update(statement)
-            .execute()
+        new Query(statement)
+            .mapResult(actual -> {
+                assertThat(actual).isSameAs(result);
+                return Mono.just(1);
+            })
             .as(StepVerifier::create)
-            .expectNext(100)
+            .expectNext(1)
             .verifyComplete();
+    }
+
+    @Test
+    void mapResultNoF() {
+        assertThatIllegalArgumentException().isThrownBy(() -> new Query(MockStatement.empty()).mapResult(null))
+            .withMessage("mappingFunction must not be null");
     }
 
 }
