+++ r2dbc/src/test/java/org/jdbi/v3/r2dbc/ResultBearingTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.r2dbc.client;
+package org.jdbi.v3.r2dbc;
 
 import java.util.function.BiFunction;
 import java.util.function.Function;
