+++ r2dbc/src/test/resources/logback-test.xml
@@ -24,7 +24,7 @@
     </appender>
 
     <logger name="com.github.jasync"   level="INFO"/>
-    <logger name="io.r2dbc.client"     level="INFO"/>
+    <logger name="org.jdbi.v3.r2dbc"   level="INFO"/>
     <logger name="io.r2dbc.postgresql" level="INFO"/>
     <logger name="io.r2dbc.mssql"      level="INFO"/>
     <logger name="org.testcontainers"  level="INFO"/>
